requirements.txt
bcrypt==3.2.0
ffmpeg==1.4
ffprobe==0.5
Flask==2.3.2
Flask-Cors==4.0.0
Flask-DotEnv==0.1.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gunicorn == 21.2.0
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
SQLAlchemy==2.0.16
pydub==0.25.1
Pyrebase4==4.7.1
python-dotenv==0.19.2
psycopg2-binary==2.9.9
flask-socketio>=5.0.0
python-socketio>=5.0.0
eventlet>=0.30.2
gevent
